refactor: Change directory structure
