+++ convert/src/labelstudio_to_badgerdoc/badgerdoc_format/annotation_converter_practic.py
@@ -1,14 +1,14 @@
 from typing import List
 
-from ..models import bd_annotation_model_practic
-from ..models.bd_annotation_model import (
+from . import bd_annotation_model_practic
+from .bd_annotation_model import (
     AnnotationLink,
     BadgerdocAnnotation,
     Obj,
     Page,
     Size,
 )
-from ..models.bd_tokens_model import Page as BadgerdocTokensPage
+from .bd_tokens_model import Page as BadgerdocTokensPage
 
 FIRST_PAGE = 0
