+++ convert/src/labelstudio_to_badgerdoc/badgerdoc_format/bd_annotation_model.py
