+++ convert/src/labelstudio_to_badgerdoc/badgerdoc_format/pdf_renderer.py
@@ -6,7 +6,7 @@ import fitz
 from fitz import Page
 
 from ...config import DEFAULT_PAGE_BORDER_OFFSET, DEFAULT_PDF_FONT_HEIGHT
-from ..models import BadgerdocToken
+from .bd_tokens_model import BadgerdocToken
 
 
 class Fonts(Enum):
