--- convert/src/labelstudio_to_badgerdoc/models/__init__.py
@@ -1,13 +1,0 @@
-from .bd_annotation_model import BadgerdocAnnotation, Obj, Page, Size
-from .bd_annotation_model_practic import DocumentLink
-from .bd_tokens_model import BadgerdocToken, Offset, Page, PageSize
-from .common import S3Path
-from .labelstudio_models import (
-    Annotation,
-    BadgerdocToLabelStudioRequest,
-    LabelStudioModel,
-    LabelStudioRequest,
-    ModelItem,
-    Prediction,
-    ResultItem,
-)
