+++ convert/src/legacy/coco_import/import_job.py
@@ -3,7 +3,7 @@ from typing import Any, Dict
 from urllib.error import HTTPError
 from uuid import uuid4
 
-from src.coco_import.import_service import import_run
+from .import_service import import_run
 from src.config import get_request_session, settings
 from src.logger import get_logger
 from src.models.coco import DataS3
